backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-genai
